Account for quotes when resolving identifiers.
We currently resolve namespaced fields (e.g. public.column_name) to a single identifier. This ensures that we resolve table metadata to include the namespace. This same logic didn't work for identifiers with quotes in them (e.g. public."column_name", because we were excluding quotes as a legal identifier character. They are legal in identifiers, so we should include those.